added common styles and side drawer
