+++ src/components/DrawerSection.tsx
@@ -1,0 +1,14 @@
+import UnfoldMoreOutlinedIcon from '@mui/icons-material/UnfoldMoreOutlined';
+
+const DrawerSection = () => {
+    return (
+        <div className='container' >
+            <div className='title section-title' >
+                <span>Finance Manager</span>
+                <UnfoldMoreOutlinedIcon className='collapse-icon' />
+            </div>
+        </div>
+    )
+}
+
+export default DrawerSection
