+++ src/pages/DashBoard.tsx
@@ -1,107 +1,19 @@
-import * as React from 'react';
+import DrawerSection from "../components/DrawerSection"
+import { Grid } from "@mui/material"
+import { Outlet } from "react-router-dom"
 
-import { AppProvider, type Navigation, type Router } from '@toolpad/core/AppProvider';
-import BarChartIcon from '@mui/icons-material/BarChart';
-import DashboardIcon from '@mui/icons-material/Dashboard';
-import { DashboardLayout } from '@toolpad/core/DashboardLayout';
-import DescriptionIcon from '@mui/icons-material/Description';
-import LayersIcon from '@mui/icons-material/Layers';
-import ShoppingCartIcon from '@mui/icons-material/ShoppingCart';
-import { createTheme } from '@mui/material/styles';
-import { Outlet } from 'react-router-dom';
-
-const NAVIGATION: Navigation = [
-  {
-    kind: 'header',
-    title: 'Main items',
-  },
-  {
-    segment: 'dashboard',
-    title: 'Dashboard',
-    icon: <DashboardIcon />,
-  },
-  {
-    segment: 'orders',
-    title: 'Orders',
-    icon: <ShoppingCartIcon />,
-  },
-  {
-    kind: 'divider',
-  },
-  {
-    kind: 'header',
-    title: 'Analytics',
-  },
-  {
-    segment: 'reports',
-    title: 'Reports',
-    icon: <BarChartIcon />,
-    children: [
-      {
-        segment: 'sales',
-        title: 'Sales',
-        icon: <DescriptionIcon />,
-      },
-      {
-        segment: 'traffic',
-        title: 'Traffic',
-        icon: <DescriptionIcon />,
-      },
-    ],
-  },
-  {
-    segment: 'integrations',
-    title: 'Integrations',
-    icon: <LayersIcon />,
-  },
-];
-
-const demoTheme = createTheme({
-  colorSchemes: { light: true, dark: true },
-  cssVariables: {
-    colorSchemeSelector: 'class',
-  },
-  breakpoints: {
-    values: {
-      xs: 0,
-      sm: 600,
-      md: 600,
-      lg: 1200,
-      xl: 1536,
-    },
-  },
-});
-
-function useDemoRouter(initialPath: string): Router {
-  const [pathname, setPathname] = React.useState(initialPath);
-
-  const router = React.useMemo(() => {
-    return {
-      pathname,
-      searchParams: new URLSearchParams(),
-      navigate: (path: string | URL) => setPathname(String(path)),
-    };
-  }, [pathname]);
-
-  return router;
-}
-
-
-const Dashboard = () => {
-
-  const router = useDemoRouter('/dashboard');
-
+const DashBoard = () => {
   return (
-    <AppProvider
-      navigation={NAVIGATION}
-      router={router}
-      theme={demoTheme}
-    >
-      <DashboardLayout>
+    <Grid container className="main-container" >
+      <Grid size={2} p={3} className="drawer-section" >
+        <DrawerSection />
+      </Grid>
+      <Grid size={10} px={2} className="content-section">
+        <div className="header" ></div>
         <Outlet />
-      </DashboardLayout>
-    </AppProvider>
-  );
+      </Grid>
+    </Grid>
+  )
 }
 
-export default Dashboard;
+export default DashBoard
