+++ src/styles/_colors.scss
@@ -10,6 +10,7 @@ $neutral-clr-100: #FFFFFF;
 
 // border color
 $border-outline: #343B4F;
+$drawer-border-outline: $neutral-clr-600;
 
 // button color
 $btn-grad: linear-gradient(128.49deg, #CB3CFF 19.86%, #7F25FB 68.34%);
