+++ src/styles/_common.scss
@@ -1,0 +1,20 @@
+.main-title,
+.section-title,
+.label-title {
+    color: $primary-txt-color;
+}
+
+.main-title {
+    font-size: 24px;
+    font-weight: 600;
+}
+
+.section-title {
+    font-size: 20px;
+    font-weight: 600;
+}
+
+.label-title {
+    font-weight: 500;
+    font-size: 16px;
+}
